+++ packages/rainbowkit/src/components/ConnectButton/ConnectButtonRenderer.tsx
@@ -1,9 +1,11 @@
 import React, { ReactNode, useContext } from 'react';
 import { useAccount, useBalance, useConfig } from 'wagmi';
+import { normalizeResponsiveValue } from '../../css/sprinkles.css';
 import { useIsMounted } from '../../hooks/useIsMounted';
 import { useMainnetEnsAvatar } from '../../hooks/useMainnetEnsAvatar';
 import { useMainnetEnsName } from '../../hooks/useMainnetEnsName';
 import { useRecentTransactions } from '../../transactions/useRecentTransactions';
+import { isMobile } from '../../utils/isMobile';
 import { useAsyncImage } from '../AsyncImage/useAsyncImage';
 import {
   AuthenticationStatus,
@@ -16,6 +18,7 @@ import {
   useModalState,
 } from '../RainbowKitProvider/ModalContext';
 import { useRainbowKitChainsById } from '../RainbowKitProvider/RainbowKitChainContext';
+import { useShowBalance } from '../RainbowKitProvider/ShowBalanceContext';
 import { ShowRecentTransactionsContext } from '../RainbowKitProvider/ShowRecentTransactionsContext';
 import { abbreviateETHBalance } from './abbreviateETHBalance';
 import { formatAddress } from './formatAddress';
@@ -62,12 +65,13 @@ export function ConnectButtonRenderer({
   const { address } = useAccount();
   const ensName = useMainnetEnsName(address);
   const ensAvatar = useMainnetEnsAvatar(ensName);
+
   const { chainId } = useAccount();
   const { chains: wagmiChains } = useConfig();
   const isCurrentChainSupported = wagmiChains.some(
     (chain) => chain.id === chainId,
   );
-  const { data: balanceData } = useBalance({ address });
+  
   const rainbowkitChainsById = useRainbowKitChainsById();
   const authenticationStatus = useAuthenticationStatus() ?? undefined;
   const rainbowKitChain = chainId ? rainbowkitChainsById[chainId] : undefined;
@@ -81,6 +85,15 @@ export function ConnectButtonRenderer({
     useRecentTransactions().some(({ status }) => status === 'pending') &&
     showRecentTransactions;
 
+  const { showBalance } = useShowBalance();
+  const shouldShowBalance = showBalance
+    ? normalizeResponsiveValue(showBalance)[
+        isMobile() ? 'smallScreen' : 'largeScreen'
+      ]
+    : false;
+  const { data: balanceData } = useBalance({
+    address: shouldShowBalance ? address : undefined,
+  });
   const displayBalance = balanceData
     ? `${abbreviateETHBalance(parseFloat(balanceData.formatted))} ${
         balanceData.symbol