+++ packages/rainbowkit/src/components/ConnectOptions/DesktopOptions.tsx
@@ -294,8 +294,8 @@ export function DesktopOptions({ onClose }: { onClose: () => void }) {
       headerBackButtonLink = connector
         ? WalletStep.Connect
         : hasExtensionAndMobile && WalletStep.Connect
-        ? initialWalletStep
-        : null;
+          ? initialWalletStep
+          : null;
       break;
     case WalletStep.Download:
       walletContent = selectedWallet && (