+++ packages/rainbowkit/src/components/ProfileDetails/ProfileDetails.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { GetEnsNameReturnType } from 'viem';
 import { GetEnsAvatarReturnType } from 'viem/actions';
-import { UseBalanceReturnType, useAccount } from 'wagmi';
-import { GetBalanceData } from 'wagmi/query';
+import { useAccount, useBalance } from 'wagmi';
 import { isMobile } from '../../utils/isMobile';
 import { Avatar } from '../Avatar/Avatar';
 import { Box } from '../Box/Box';
@@ -21,7 +20,6 @@ import { ProfileDetailsAction } from './ProfileDetailsAction';
 
 interface ProfileDetailsProps {
   address: ReturnType<typeof useAccount>['address'];
-  balanceData: UseBalanceReturnType<GetBalanceData>['data'];
   ensAvatar: GetEnsAvatarReturnType | undefined;
   ensName: GetEnsNameReturnType | undefined;
   onClose: () => void;
@@ -30,17 +28,18 @@ interface ProfileDetailsProps {
 
 export function ProfileDetails({
   address,
-  balanceData,
   ensAvatar,
   ensName,
   onClose,
   onDisconnect,
 }: ProfileDetailsProps) {
   const showRecentTransactions = useContext(ShowRecentTransactionsContext);
+
+  const { data: balanceData } = useBalance({
+    address,
+  });
+
   const [copiedAddress, setCopiedAddress] = useState(false);
-
-  const { i18n } = useContext(I18nContext);
-
   const copyAddressAction = useCallback(() => {
     if (address) {
       navigator.clipboard.writeText(address);
@@ -68,6 +67,8 @@ export function ProfileDetails({
     : undefined;
   const titleId = 'rk_profile_title';
   const mobile = isMobile();
+
+  const { i18n } = useContext(I18nContext);
 
   return (
     <>