+++ packages/rainbowkit/src/wallets/walletConnectors/metaMaskWallet/metaMaskWallet.ts
@@ -8,7 +8,7 @@ export interface MetaMaskWalletOptions {
   projectId: string;
 }
 
-function isMetaMask(ethereum?: typeof window['ethereum']): boolean {
+function isMetaMask(ethereum?: (typeof window)['ethereum']): boolean {
   // Logic borrowed from wagmi's MetaMaskConnector
   // https://github.com/wagmi-dev/references/blob/main/packages/connectors/src/metaMask.ts
   if (!ethereum?.isMetaMask) return false;