+++ site/next.config.js
@@ -22,6 +22,7 @@ const nextConfig = {
       'ru',
       'th',
       'tr',
+      'ua',
       'zh-CN',
     ],
     defaultLocale: 'en-US',