+++ examples/with-next-siwe-iron-session/CHANGELOG.md
@@ -1,5 +1,14 @@
 # with-next-siwe-iron-session
 
+## 0.0.58
+
+### Patch Changes
+
+- Updated dependencies [24b5a88]
+- Updated dependencies [7565fb2]
+- Updated dependencies [5a184e9]
+  - @rainbow-me/rainbowkit@1.3.3
+
 ## 0.0.57
 
 ### Patch Changes