+++ packages/create-rainbowkit/templates/next-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # rainbowkit-next-app
 
+## 0.0.78
+
+### Patch Changes
+
+- Updated dependencies [24b5a88]
+- Updated dependencies [7565fb2]
+- Updated dependencies [5a184e9]
+  - @rainbow-me/rainbowkit@1.3.3
+
 ## 0.0.77
 
 ### Patch Changes